feat(ses): Add minimal Compartment to SES-lite lockdown layer
